+++ projects/core/src/main/kotlin/promises-api.kt
@@ -335,9 +335,13 @@ fun <V, R> Promise<V, Exception>.then(context: Context, bind: (V) -> R): Promise
  *
  * @param bind the transform function.
  */
-infix inline fun <V, R> Promise<V, Exception>.thenUse(
+infix inline fun <V, R> Promise<V, Exception>.thenApply(
         crossinline bind: V.() -> R): Promise<R, Exception> = then { it.bind() }
 
+
+@Deprecated("renamed to 'thenApply'", ReplaceWith("thenApply(bind)"))
+infix inline fun <V, R> Promise<V, Exception>.thenUse(
+        crossinline bind: V.() -> R): Promise<R, Exception> = then { it.bind() }
 
 /**
  * Transforms any `Promise<V, E>` into a Promise<Unit, Unit>.
@@ -395,5 +399,36 @@ fun <V, E> Promise<V, E>.toSuccessVoid(context: Context = this.context): Promise
     val deferred = deferred<Unit, E>(context)
     success { deferred.resolve() }
     fail { deferred.reject(it) }
+    return deferred.promise
+}
+
+
+/**
+ * Unwraps any nested Promise.
+ *
+ * Unwraps any nested Promise. By default the returned `Promise` will operate on the same `Context` as its parent
+ * `Promise`, no matter what the `Context` of the nested `Promise` is. If you want the resulting promise to operate on
+ * a different `Context` you can provide one.
+ *
+ * Function tries to be as efficient as possible in cases where this or the nested `Promise` is already resolved. This
+ * means that this function might or might not create a new `Promise`, it all depends on the current state.
+ *
+ * @param context the `Context` on which the returned `Promise` operates
+ * @return the unwrapped Promise
+ */
+fun <V, E> Promise<Promise<V, E>, E>.unwrap(context: Context = this.context): Promise<V, E> {
+    if (isDone()) when {
+        isSuccess() -> return get().withContext(context)
+        isFailure() -> return Promise.ofFail(getError(), context)
+    }
+
+    val deferred = deferred<V, E>(context)
+    success {
+        nested ->
+        nested success { deferred resolve it }
+        nested fail { deferred reject it }
+    } fail {
+        deferred reject it
+    }
     return deferred.promise
 }